program now allow quantity to be added and removed by 0.1 when price type = kg, program also displays correct total with this change

+++ code rebase/Kiwi Basket - (App Development)/db.py
@@ -44,7 +44,7 @@ cursor.execute(sql_query)
 sql_query = """CREATE TABLE IF NOT EXISTS ListsItems (
         lid integer NOT NULL, 
         iid integer NOT NULL, 
-        lquantity integer NOT NULL, 
+        lquantity float NOT NULL, 
         primary key (lid, iid), 
         foreign key (lid) references Lists(lid), 
         foreign key (iid) references Items(iid)
